Move files
Moved files from the test_models directory
to the tests directory.
The unittest method could not discover tests in the
test_models directory, hence the change.
This also conforms to the alx checker that checks for the tests.

The program to check for the tests executes the following in a
subprocess:
	"Python3 -m unittest discover tests"
This fails to discover tests in the tests folder which are not
Python files